build complete with addressing some warnings

+++ src/integrator/sycl/collocation_masked_combined_kernels.hpp
@@ -21,12 +21,12 @@ using namespace GauXC::sycl;
                                                    XCTaskDevice<T>* device_tasks,
                                                    cl::sycl::nd_item<3> item_ct) {
 
-        const int tid_x = item_ct.get_group(2) * item_ct.get_local_range().get(2) +
+        const size_t tid_x = item_ct.get_group(2) * item_ct.get_local_range().get(2) +
             item_ct.get_local_id(2);
-        const int tid_y = item_ct.get_group(1) * item_ct.get_local_range().get(1) +
+        const size_t tid_y = item_ct.get_group(1) * item_ct.get_local_range().get(1) +
             item_ct.get_local_id(1);
 
-        const int batch_id = item_ct.get_group(0);
+        const size_t batch_id = item_ct.get_group(0);
 
         if( batch_id < ntasks ) {
 
@@ -88,12 +88,12 @@ using namespace GauXC::sycl;
         XCTaskDevice<T>* device_tasks,
         cl::sycl::nd_item<3> item_ct) {
 
-        const int tid_x = item_ct.get_group(2) * item_ct.get_local_range().get(2) +
+        const size_t tid_x = item_ct.get_group(2) * item_ct.get_local_range().get(2) +
             item_ct.get_local_id(2);
-        const int tid_y = item_ct.get_group(1) * item_ct.get_local_range().get(1) +
+        const size_t tid_y = item_ct.get_group(1) * item_ct.get_local_range().get(1) +
             item_ct.get_local_id(1);
 
-        const int batch_id = item_ct.get_group(0);
+        const size_t batch_id = item_ct.get_group(0);
 
         if( batch_id < ntasks ) {
